Made playlist screen where user can add new playlists

+++ screens/playlists.js
@@ -1,5 +1,7 @@
 import React from 'react';
-import { View, Text, StyleSheet, TouchableOpacity, AsyncStorage } from 'react-native';
+import { Dimensions, View, Text, StyleSheet, TouchableOpacity, AsyncStorage, FlatList, Image, Modal, TextInput } from 'react-native';
+
+const { width: DEVICE_WIDTH, height: DEVICE_HEIGHT } = Dimensions.get('window');
 
 export default class Playlists extends React.Component {
     constructor() {
@@ -7,32 +9,53 @@ export default class Playlists extends React.Component {
 
         this.state= {
             allPlaylists: [],
+            modalVisible: false,
+            playlistName: '',
+            newPlaylistError: '',
         }
     }
 
     async componentDidMount() {
         const Playlists = await AsyncStorage.getItem('playlists');
-        await this.setState({allPlaylists: JSON.parse(Playlists)});
+
+        if (Playlists != null) {
+            await this.setState({allPlaylists: JSON.parse(Playlists)});
+        }  
+    }
+
+    componentWillUnmount() {
+        this.setState({allPlaylists: []});
     }
 
     createPlaylist = async () => {
-        const playlistsToSave = { 'name': 'More Music'}
-        const existingProducts = await AsyncStorage.getItem('playlists')
+        if (this.state.playlistName !== '') {
+            const playlistsToSave = { 'name': this.state.playlistName}
+            const existingProducts = await AsyncStorage.getItem('playlists')
 
-        let newPlaylists = JSON.parse(existingProducts);
-        if( !newPlaylists ){
-            newPlaylists = []
+            if (existingProducts !== null && existingProducts.includes(this.state.playlistName)) {
+                this.setState({newPlaylistError: 'Name is already taken'})
+            } else {
+
+                let newPlaylists = JSON.parse(existingProducts);
+                if( !newPlaylists ){
+                    newPlaylists = []
+                }
+
+                newPlaylists.push( playlistsToSave )
+
+                await AsyncStorage.setItem('playlists', JSON.stringify(newPlaylists))
+                    .then(() => {
+                        this.props.navigation.replace('Playlists');
+                    })
+                    .catch((e) => {
+                        console.log(e);
+                    })
+
+            }
+        } else {
+            this.setState({newPlaylistError: 'Please enter a name for your playist'})
         }
-
-        newPlaylists.push( playlistsToSave )
-
-        await AsyncStorage.setItem('playlists', JSON.stringify(newPlaylists))
-            .then(() => {
-                this.props.navigation.replace('Playlists');
-            })
-            .catch((e) => {
-                console.log(e);
-            })
+        
     }
 
     render() {
@@ -40,18 +63,58 @@ export default class Playlists extends React.Component {
             <View style={styles.container}>
     
                 <View style={styles.buttonContainer}>
-                    <TouchableOpacity style={styles.playlistButton} onPress={this.createPlaylist}>
+                    <TouchableOpacity style={styles.playlistButton} onPress={() => {this.setState({modalVisible: true})}}>
                         <Text style={styles.playlistBtnText}>Create playlist</Text>
                     </TouchableOpacity>
                 </View>
-    
-                <Text>Playlists screen</Text>
 
-                <View>
-                    {this.state.allPlaylists.map(item => {
-                        return <Text>{item.name}</Text>
-                    })}
-                </View>
+                <FlatList
+                    numColumns={2}
+                    data={this.state.allPlaylists}
+                    keyExtractor={(item, index) => index.toString()}
+                    renderItem={({item}) =>
+                        <TouchableOpacity  key={Math.random()} style={styles.playlistContainer}>
+                            <Image style={styles.playlistImage} source={require('../assets/music-icon.png')} />
+                            <Text style={styles.playlistText}>{item.name}</Text>
+                        </TouchableOpacity>
+
+                    }
+                /> 
+
+                <Modal
+                    style={styles.playlistModal}
+                    animationType='slide'
+                    visible={this.state.modalVisible}
+                    transparent={true}
+                >
+                    <View style={styles.modalBackContainer}>
+                        <View style={styles.modalContainer}>
+                        
+                        <View>
+                            <Text>New playlist name:</Text>
+
+                            <TextInput 
+                                style={styles.playlistInput}
+                                onChangeText={(text) => {this.setState({playlistName: text, newPlaylistError: ''})}}
+                                // value={this.state.playlistName}
+                            />
+
+                            <Text style={styles.playlistErrorText}>{this.state.newPlaylistError}</Text>
+                        </View>
+
+                        <View style={styles.playlistButtonsContainer}>
+                            <TouchableOpacity style={styles.buttonCancel} onPress={() => {this.setState({modalVisible: false})}}>
+                                <Text style={styles.buttonText}>Cancel</Text>
+                            </TouchableOpacity>
+
+                            <TouchableOpacity style={styles.buttonSave} onPress={this.createPlaylist}>
+                                <Text  style={styles.buttonText}>Save playlist</Text>
+                            </TouchableOpacity>
+                        </View>
+                            
+                        </View>
+                    </View>
+                </Modal>
 
             </View>
         )
@@ -60,12 +123,12 @@ export default class Playlists extends React.Component {
 
 const styles = StyleSheet.create({
     container: {
-      flex: 1,
+    flex: 1,
       backgroundColor: '#ccfffd',
+      alignItems: 'center',
     },
     buttonContainer: {
         padding: 30,
-        alignItems: 'flex-end'
     },
     playlistButton: {
         width: 140,
@@ -76,5 +139,60 @@ const styles = StyleSheet.create({
     playlistBtnText: {
         fontSize: 18,
         fontWeight: 'bold'
+    },
+    playlistContainer: {
+        margin: '5%',
+        alignItems: 'center',
+        width: '40%',
+        height: '80%',
+    },
+    playlistImage: {
+        height: 100,
+        width: 130,
+    },
+    playlistText: {
+        fontWeight: 'bold',
+        fontSize: 16,
+    },
+    modalBackContainer: {
+        backgroundColor: '#000000aa',
+        flex: 1, 
+    },
+    modalContainer: {
+        alignSelf: 'center',
+        marginVertical: '35%',
+        backgroundColor: '#fff' ,
+        minHeight: DEVICE_HEIGHT / 3.5,
+        maxHeight: DEVICE_HEIGHT / 3.5,
+        width: '70%',
+        padding: 20, 
+        borderRadius: 10,
+        flex: 1, 
+    },
+    playlistButtonsContainer: {
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+    },
+    playlistInput: {
+        marginVertical: '5%',
+        borderWidth: 1,
+    },
+    buttonCancel: {
+        padding: 10,
+        backgroundColor: 'darkgrey',
+        borderRadius: 10,
+    },
+    buttonSave: {
+        padding: 10,
+        backgroundColor: 'lightblue',
+        borderRadius: 10,
+    },
+    buttonText: {
+        fontSize: 16,
+        fontWeight: 'bold',
+    },
+    playlistErrorText: {
+        color: 'red',
+        marginVertical: '5%',
     }
   });
